+++ crates/rynix-rir/src/lower.rs
@@ -108,6 +108,7 @@ fn map_ty(analysis: &Analysis, ty: TypeId) -> IrTy {
         | TypeKind::Vec
         | TypeKind::VecStr
         | TypeKind::Map
+        | TypeKind::MapStrI64
         | TypeKind::Slice(_)
         | TypeKind::Struct(_)
         | TypeKind::Fn { .. } => IrTy::Ptr,
@@ -4757,12 +4758,15 @@ impl LowerCtx<'_, '_> {
         }
         let soft = match (method.as_str(), kind) {
             ("insert", Some(TypeKind::Map)) => "map_insert",
+            ("insert", Some(TypeKind::MapStrI64)) => "map_str_i64_insert",
             ("push", Some(TypeKind::Vec)) => "vec_push",
             ("push", Some(TypeKind::VecStr)) => "vec_str_push",
             ("get", Some(TypeKind::Map)) => "map_get",
+            ("get", Some(TypeKind::MapStrI64)) => "map_str_i64_get",
             ("get", Some(TypeKind::Vec)) => "vec_get",
             ("get", Some(TypeKind::VecStr)) => "vec_str_get",
             ("len", Some(TypeKind::Map)) => "map_len",
+            ("len", Some(TypeKind::MapStrI64)) => "map_str_i64_len",
             ("len", Some(TypeKind::Vec)) => "vec_len",
             ("len", Some(TypeKind::VecStr)) => "vec_str_len",
             (other, _) => other,
@@ -4797,6 +4801,12 @@ impl LowerCtx<'_, '_> {
             "map_insert" => (self.interner.intern("rynix_rt_map_i64_insert"), IrTy::Unit),
             "map_get" => (self.interner.intern("rynix_rt_map_i64_get"), IrTy::I64),
             "map_len" => (self.interner.intern("rynix_rt_map_i64_len"), IrTy::I64),
+            "map_str_i64_new" => (self.interner.intern("rynix_rt_map_str_i64_new"), IrTy::Ptr),
+            "map_str_i64_insert" => {
+                (self.interner.intern("rynix_rt_map_str_i64_insert"), IrTy::Unit)
+            }
+            "map_str_i64_get" => (self.interner.intern("rynix_rt_map_str_i64_get"), IrTy::I64),
+            "map_str_i64_len" => (self.interner.intern("rynix_rt_map_str_i64_len"), IrTy::I64),
             "tcp_listen" => (self.interner.intern("rynix_rt_tcp_listen"), IrTy::I64),
             "tcp_accept" => (self.interner.intern("rynix_rt_tcp_accept"), IrTy::I64),
             "tcp_connect" => (self.interner.intern("rynix_rt_tcp_connect"), IrTy::I64),