+++ crates/rynix-sema/src/ty.rs
@@ -36,6 +36,8 @@ pub enum TypeKind {
     VecStr,
     /// Region Map[i64, i64] handle (ADR-0006).
     Map,
+    /// Region Map[str, i64] handle (ADR-0017).
+    MapStrI64,
     Slice(TypeId),
     Fn {
         params: Vec<TypeId>,
@@ -65,6 +67,7 @@ pub struct TypeCtx {
     pub ty_vec: TypeId,
     pub ty_vec_str: TypeId,
     pub ty_map: TypeId,
+    pub ty_map_str_i64: TypeId,
     pub ty_module: TypeId,
 }
 
@@ -91,6 +94,7 @@ impl TypeCtx {
             ty_vec: TypeId(0),
             ty_vec_str: TypeId(0),
             ty_map: TypeId(0),
+            ty_map_str_i64: TypeId(0),
             ty_module: TypeId(0),
         };
         ctx.ty_error = ctx.intern(TypeKind::Error);
@@ -105,6 +109,7 @@ impl TypeCtx {
         ctx.ty_vec = ctx.intern(TypeKind::Vec);
         ctx.ty_vec_str = ctx.intern(TypeKind::VecStr);
         ctx.ty_map = ctx.intern(TypeKind::Map);
+        ctx.ty_map_str_i64 = ctx.intern(TypeKind::MapStrI64);
         ctx.ty_module = ctx.intern(TypeKind::Module);
         ctx
     }
@@ -186,6 +191,7 @@ impl TypeCtx {
             TypeKind::Vec => "Vec[i64]".into(),
             TypeKind::VecStr => "Vec[str]".into(),
             TypeKind::Map => "Map[i64, i64]".into(),
+            TypeKind::MapStrI64 => "Map[str, i64]".into(),
             TypeKind::Module => "<module>".into(),
             TypeKind::Slice(e) => format!("[{}]", self.display(*e, resolve_name, interner)),
             TypeKind::Fn { params, ret } => {